requirements.txt
matplotlib~=3.2.2
numpy~=1.19.1
natsort~=7.0.1
pillow~=7.2.0
torchvision~=0.7.0
setuptools~=49.2.0